requirements.txt
requests
openai
pandas
PyYAML
python-docx
schedule
smtplib
